requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
openai
python-dotenv
pydantic